exceptions implemented

+++ src/main/java/com/project/cscb869/services/implementations/AutoServiceImplementation.java
@@ -5,6 +5,7 @@ import com.project.cscb869.data.entity.Car;
 import com.project.cscb869.data.entity.Worker;
 import com.project.cscb869.data.repository.AutoServiceRepository;
 import com.project.cscb869.data.repository.WorkerRepository;
+import com.project.cscb869.exceptions.NotFoundException;
 import com.project.cscb869.services.AutoServiceService;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -26,9 +27,8 @@ public class AutoServiceImplementation  implements AutoServiceService {
     @Override
     public AutoService getService(long id) {
         return autoServiceRepository.findById(id)
-                .orElseThrow(() -> new IllegalArgumentException("Invalid auto service id: " + id));
+                .orElseThrow(() -> new NotFoundException("Auto Service with id "+ id +" doesn't exist!"));
     }
-
     @Override
     public AutoService createService(AutoService autoService) {
         return autoServiceRepository.save(autoService);
@@ -36,17 +36,26 @@ public class AutoServiceImplementation  implements AutoServiceService {
 
     @Override
     public AutoService updateService(long id, AutoService autoService) {
-        autoService.setId(id);
-        return autoServiceRepository.save(autoService);
+        if(autoServiceRepository.findById(id).isPresent()){
+            autoService.setId(id);
+            return autoServiceRepository.save(autoService);
+        }else {
+            throw new NotFoundException("Auto Service with id "+ id +" doesn't exist!");
+        }
     }
-
     @Override
     public void deleteService(long id) {
-        autoServiceRepository.deleteById(id);
+        if(autoServiceRepository.findById(id).isPresent()){
+            autoServiceRepository.deleteById(id);
+        }else{
+            throw new NotFoundException("Auto Service with id "+ id +" doesn't exist!");
+        }
     }
-
     @Override
     public AutoService getAutoServiceByName(String name){
+        if(autoServiceRepository.getAutoServiceByName(name) == null){
+            throw new NotFoundException("Auto Service with name "+ name +" doesn't exist!");
+        }
         return autoServiceRepository.getAutoServiceByName(name);
     }
     @Override
@@ -55,7 +64,7 @@ public class AutoServiceImplementation  implements AutoServiceService {
     }
     @Override
     public AutoService addWorker(long serviceId, Worker worker){
-        AutoService autoService = autoServiceRepository.findById(serviceId).orElseThrow(() -> new IllegalArgumentException("Invalid auto service id: " + serviceId));
+        AutoService autoService = autoServiceRepository.findById(serviceId).orElseThrow(() -> new NotFoundException("Auto Service with id "+ serviceId +" doesn't exist!"));
 
         List<Worker> newWorkers = autoService.getWorkers();
         newWorkers.add(worker);
